requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
multidict==3.3.0
praw==5.1.0
prawcore==0.12.0
pyaml==17.10.0
PyMySQL==0.7.11
PyYAML==3.12
requests==2.18.4
SQLAlchemy==1.1.14
update-checker==0.16
urllib3==1.22
Werkzeug==0.12.2
yarl==0.13.0
